archives: disable record_timestamps per model instead of globally
This probably messed (due to multithreading or something) with the
creation of records in other models where created_at is not explicitly
set but required to be not null (i.e. they're relying on rails setting
it)

What happened for instance is that I would get null errors on the
created_at field when uploading a image (on the posts table) since
created_at is required to be not null there and since rails is normally
reponsible for setting those fields the only explanation was that it was
disabled somewhere and since I made this change it hasn't happened
again.

https://en.wikibooks.org/wiki/Ruby_on_Rails/ActiveRecord/Timestamps